Change from matplotlib notebook to inline
If we are recommending Jupyter Lab instead of Jupyter Notebook,
then the %matplotlib notebook widget doesn't work.
See this stackoverflow answer for suggested things
to install so it can work,
https://stackoverflow.com/questions/51922480/javascript-error-ipython-is-not-defined-in-jupyterlab
but an easier way to avoid the problem is
to use matploltilb inline.
Won't be interactive and zoomable, but at least will work.

I did a search/replace for all the notebooks.