Equilibrium: Work around NUIG incompatibilities with Cantera 2.6
For the equilibrium calculations using the NUIG mechanism,
unfortunately the ck2yaml doesn't make a valid mechanism,
and we have to faff around fixing it.

Perhaps we should just edit the chemkin files, as the point
of this notebook is really to discuss equilibrium, but this 
version at least shows how the sausage is made.